Add setup_memery_cache! method
This method isn't strictly necessary even for its _own goal_, since
calling `clear_memery_cache!` in the initializer would work just as
well. But a well-named method will be less confusing.

If there were a way to do this _automatically_ it would be better, but I
couldn't come up with anything that wasn't extremely dirty, and every
trick I thought of had edges one could run into when I thought further.
Better to be safe, and leave it opt-in and simple.